Move `blake2b` and `next_address` tests to `runtime.rs`
