+++ execution_engine_testing/tests/src/test/contract_api/mod.rs
@@ -1,5 +1,4 @@
 mod account;
-mod blake2b;
 mod create_purse;
 mod dictionary;
 mod get_arg;
@@ -13,8 +12,8 @@ mod main_purse;
 mod mint_purse;
 mod multisig_authorization;
 mod named_dictionaries;
-mod next_address;
 mod revert;
+mod runtime;
 mod subcall;
 mod transfer;
 mod transfer_cached;
