+++ execution_engine_testing/tests/src/test/contract_api/runtime.rs
@@ -1,24 +1,26 @@
+use std::collections::HashSet;
+
 use rand::Rng;
 
 use casper_engine_test_support::{
     ExecuteRequestBuilder, InMemoryWasmTestBuilder, DEFAULT_ACCOUNT_ADDR,
     PRODUCTION_RUN_GENESIS_REQUEST,
 };
-use casper_types::{crypto, runtime_args, RuntimeArgs, BLAKE2B_DIGEST_LENGTH};
+use casper_types::{crypto, runtime_args, RuntimeArgs, BLAKE2B_DIGEST_LENGTH, KEY_HASH_LENGTH};
 
 const BLAKE2B_WASM: &str = "blake2b.wasm";
 const ARG_BYTES: &str = "bytes";
 const HASH_RESULT: &str = "hash_result";
 
-fn get_digest(builder: &InMemoryWasmTestBuilder) -> [u8; BLAKE2B_DIGEST_LENGTH] {
+const NEXT_ADDRESS_WASM: &str = "next_address.wasm";
+const NEXT_ADDRESS_RESULT: &str = "next_address_result";
+
+fn get_value<const COUNT: usize>(builder: &InMemoryWasmTestBuilder, result: &str) -> [u8; COUNT] {
     let account = builder
         .get_account(*DEFAULT_ACCOUNT_ADDR)
         .expect("should have account");
 
-    let uref = account
-        .named_keys()
-        .get(HASH_RESULT)
-        .expect("should have value");
+    let uref = account.named_keys().get(result).expect("should have value");
 
     builder
         .query(None, *uref, &[])
@@ -28,6 +30,34 @@ fn get_digest(builder: &InMemoryWasmTestBuilder) -> [u8; BLAKE2B_DIGEST_LENGTH] 
         .expect("should be CLValue")
         .into_t()
         .expect("should convert")
+}
+
+#[ignore]
+#[test]
+fn should_return_distinct_next_addresses() {
+    const RUNS: usize = 10;
+
+    let mut builder = InMemoryWasmTestBuilder::default();
+
+    builder.run_genesis(&PRODUCTION_RUN_GENESIS_REQUEST);
+
+    let all_addresses: HashSet<_> = (0..RUNS)
+        .map(|_| {
+            let exec_request = ExecuteRequestBuilder::standard(
+                *DEFAULT_ACCOUNT_ADDR,
+                NEXT_ADDRESS_WASM,
+                runtime_args! {},
+            )
+            .build();
+
+            builder.exec(exec_request).commit().expect_success();
+
+            get_value::<KEY_HASH_LENGTH>(&builder, NEXT_ADDRESS_RESULT)
+        })
+        .collect();
+
+    // Assert that each address is unique.
+    assert_eq!(all_addresses.len(), RUNS)
 }
 
 #[ignore]
@@ -55,7 +85,7 @@ fn should_hash() {
 
         builder.exec(exec_request).commit().expect_success();
 
-        let digest = get_digest(&builder);
+        let digest = get_value::<BLAKE2B_DIGEST_LENGTH>(&builder, HASH_RESULT);
         let expected_digest = crypto::blake2b(&input);
         assert_eq!(digest, expected_digest);
     }
